backend/requirements.txt
# FastAPI Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Redis
redis==5.2.0
hiredis==3.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
bcrypt==4.2.1

# HTTP Client
httpx==0.28.1
requests==2.32.3

# Date & Time
python-dateutil==2.9.0.post0

# Email
emails==0.6

# Utilities
python-dotenv==1.0.1
email-validator==2.2.0

# Monitoring & Logging
loguru==0.7.3

# Excel/CSV Export
openpyxl==3.1.5
pandas==2.2.3

# PDF Generation
reportlab==4.2.5

# Background Tasks (optional)
celery==5.4.0

# Testing (development)
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
